revert "Move helper functions to resize A3 cluster as GPU util"
