+++ dags/multipod/maxtext_gpu_end_to_end.py
@@ -16,20 +16,110 @@
 
 
 import datetime
+import tempfile
 
 from airflow import models
+from airflow.decorators import task
+from airflow.hooks.subprocess import SubprocessHook
 from airflow.utils.task_group import TaskGroup
 from dags import composer_env
 from dags.common import test_owner
-from dags.common.vm_resource import XpkClusters, DockerImage
+from dags.common.vm_resource import XpkClusters, CpuVersion, DockerImage, GpuVersion, Project, TpuVersion, Zone
 from dags.multipod.configs import gke_config
-from xlml.utils.gpu import scale_up_a3_cluster, scale_down_a3_cluster
+from xlml.utils import gke
 
 # Run once a day at 4 am UTC (8 pm PST)
 SCHEDULED_TIME = "0 4 * * *" if composer_env.is_prod_env() else None
 
 # Number of nodes on A3 cluster to be scaled up to
 A3_NUM_NODES = 10
+
+
+def configure_project_and_cluster(project: str, cluster_name: str, zone: str):
+  region = gke.zone_to_region(zone)
+
+  gcloud_command = (
+      f"gcloud config set project {project}",
+      "sudo chown -R airflow:airflow /home/airflow/composer_kube_config",
+      f"gcloud container clusters get-credentials {cluster_name}"
+      f"  --region {region}",
+  )
+  return gcloud_command
+
+
+def resize_a3_cluster(cluster_name: str, zone: str, num_nodes: int):
+  region = gke.zone_to_region(zone)
+  node_pool = f"{cluster_name}-np-0"
+
+  gcloud_command = (
+      f"gcloud container clusters resize {cluster_name}"
+      f"  --quiet --region {region}"
+      f"  --node-pool {node_pool}"
+      f"  --num-nodes {num_nodes}",
+  )
+  return gcloud_command
+
+
+def wait_for_cluster_ready():
+  kubectl_command = (
+      "kubectl wait --for=condition=Ready nodes --all --timeout=5m",
+  )
+  return kubectl_command
+
+
+@task
+def scale_up_a3_cluster():
+  with tempfile.TemporaryDirectory() as tmpdir:
+    hook = SubprocessHook()
+
+    result = hook.run_command(
+        [
+            "bash",
+            "-c",
+            ";".join(
+                configure_project_and_cluster(
+                    Project.SUPERCOMPUTER_TESTING.value,
+                    XpkClusters.GPU_A3_CLUSTER.name,
+                    XpkClusters.GPU_A3_CLUSTER.zone,
+                )
+                + resize_a3_cluster(
+                    XpkClusters.GPU_A3_CLUSTER.name,
+                    XpkClusters.GPU_A3_CLUSTER.zone,
+                    A3_NUM_NODES,
+                )
+                + wait_for_cluster_ready()
+            ),
+        ],
+        cwd=tmpdir,
+    )
+    assert result.exit_code == 0, f"Command failed with code {result.exit_code}"
+
+
+@task
+def scale_down_a3_cluster():
+  with tempfile.TemporaryDirectory() as tmpdir:
+    hook = SubprocessHook()
+
+    result = hook.run_command(
+        [
+            "bash",
+            "-c",
+            ";".join(
+                configure_project_and_cluster(
+                    Project.SUPERCOMPUTER_TESTING.value,
+                    XpkClusters.GPU_A3_CLUSTER.name,
+                    XpkClusters.GPU_A3_CLUSTER.zone,
+                )
+                + resize_a3_cluster(
+                    XpkClusters.GPU_A3_CLUSTER.name,
+                    XpkClusters.GPU_A3_CLUSTER.zone,
+                    0,
+                )
+            ),
+        ],
+        cwd=tmpdir,
+    )
+    assert result.exit_code == 0, f"Command failed with code {result.exit_code}"
 
 
 def run_maxtext_tests(dag: models.DAG):
@@ -157,7 +247,7 @@ with models.DAG(
     catchup=False,
 ) as dag:
   with TaskGroup(group_id="scale_up", dag=dag) as scale_up:
-    scale_up_a3_cluster(A3_NUM_NODES)
+    scale_up_a3_cluster()
 
   with TaskGroup(
       group_id="run_tests", dag=dag, prefix_group_id=False
